Use cookie to display explorer connect error in OperatorUI
