+++ operator_ui/package.json
@@ -35,6 +35,7 @@
     "bignumber.js": "^8.0.1",
     "change-case": "^3.0.2",
     "console.table": "^0.10.0",
+    "cookie": "^0.4.0",
     "core-js": "^3.0.0",
     "cross-env": "^5.2.0",
     "enzyme": "^3.4.4",
@@ -77,5 +78,8 @@
     "url": "^0.11.0",
     "use-react-hooks": "^1.0.7",
     "uuid": "^3.3.2"
+  },
+  "devDependencies": {
+    "@types/cookie": "^0.3.3"
   }
 }
