+++ operator_ui/src/components/Notifications/UnhandledError.tsx
@@ -4,9 +4,9 @@ import { Link } from 'react-router-dom'
 const UnhandledError = () => (
   <React.Fragment>
     Unhandled error. Please help us by opening a{' '}
-    <Link to="https://github.com/smartcontractkit/chainlink/issues/new">
+    <a href="https://github.com/smartcontractkit/chainlink/issues/new">
       bug report
-    </Link>
+    </a>
   </React.Fragment>
 )
